Fix indentation issue in multiline expressions
For multiline expressions there can be no additional whitespaces in subsequent lines.
They should be aligned to first line of the expression.

Example:
```
\t\tthis.isMultiline()
\t\t\t.expression()
```
For such input expression the replacement text must be like this:
```
this.isMultiline()
\t.expression()
```
In other words all subsequent lines leading whitespaces has to be aligned like the first line of the expression started at the beginning of the line.